to_do_app/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-crispy-forms==1.9.2
djangorestframework==3.12.2
idna==2.10
install==1.3.3
numpy==1.19.0
Pillow==7.2.0
pygame==1.9.6
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
